pyproject.toml
[project]
name = "production-env"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "argparse>=1.4.0",
    "arrow>=1.3.0",
    "credit>=0.0.1",
    "dask>=2025.9.0",
    "datetime>=5.5",
    "hyperopt>=0.2.7",
    "ipykernel>=6.30.1",
    "mlflow>=3.3.2",
    "numpy>=2.3.3",
    "shap>=0.47.0",
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "ucimlrepo>=0.0.7",
    "utils>=1.0.2",
    "yfinance>=0.2.65",
]
